fusion3-common: Fix radio when device encrypted
For a while, Fusion 3 devices have not worked properly with encryption - the device would
boot, but there would be no radio signal. Consultations with Sony indicated that this
would be caused by init services, which were not being correctly started after the
device unlocked the data partition. After enabling qmuxd and qmiclient, and setting them
to late_start, this issue was resolved, and the RIL would initialise and operate.

To fix data connection, netmgrd needs to also be late_start.

Change-Id: I3e9472318964d767f93df3c8aba9e95f0cec00ac